+++ src/Maverick/Animation/Hitbox/HitEllipse/HitCone/HitCone.hpp
@@ -5,10 +5,10 @@
 
 class HitCone : public HitEllipse{
 	public:
-		HitCone(float* x_base, float* y_base, float x_offset, float y_offset, float x_radius, float y_radius, float angle, float slice_prop, unsigned int type);
+		HitCone(double* x_base, double* y_base, double x_offset, double y_offset, double x_radius, double y_radius, float angle, float slice_prop, unsigned int type);
 		~HitCone();
 
-		void setScale(float x_scale, float y_scale);
+		void setScale(double x_scale, double y_scale);
 
 		float getAngle();
 		float getProp();
