+++ src/Maverick/Animation/Hitbox/HitRect/HitRect.hpp
@@ -5,7 +5,7 @@
 
 class HitRect : public Hitbox{
 	public:
-		HitRect(float* x_base, float* y_base, float x_offset, float y_offset, float width, float height, unsigned int type);
+		HitRect(double* x_base, double* y_base, double x_offset, double y_offset, float width, float height, unsigned int type);
 		~HitRect();
 
 		float getWidth();
@@ -17,8 +17,8 @@ class HitRect : public Hitbox{
         float getBotBound() override;
 		float getDrawAxis() override;
 
-		void setScale(float x_scale, float y_scale) override;
-		bool isPointInside(float x_coord, float y_coord) override;
+		void setScale(double x_scale, double y_scale) override;
+		bool isPointInside(double x_coord, double y_coord) override;
 
 		bool checkCollision(Hitbox* other) override;
